datafly_visualizer/requirements.txt
Flask==2.2.2
Flask-SocketIO==5.4.1
scikit-learn==1.4.1.post1
umap-learn==0.5.7
pandas==2.0.3
numpy==1.24.3
matplotlib==3.9.3
Werkzeug==2.2.2